+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All changes to ReportedIP Hive are documented here.
 
-## [Unreleased]
+## [2.1.20] — 2026-06-25
 
 ### Fixed